+++ src/shared/enum.ts
@@ -1,0 +1,11 @@
+
+export enum CardBrand {
+  AMERICAN_EXPRESS = 'American Express',
+  DISCOVER = 'Discover',
+  JCB = 'JCB',
+  DINERS_CLUB = 'Diners Club',
+  VISA = 'Visa',
+  MASTERCARD = 'MasterCard',
+  UNIONPAY = 'UnionPay',
+  UNKNOWN = 'Unknown',
+}
