+++ src/shared/index.ts
@@ -1,15 +1,9 @@
+import type {CardBrand} from './enum';
+
+export * from './enum';
+
 export type AvailabilityResponse = { available: boolean };
 
-export enum CardBrand {
-  AMERICAN_EXPRESS = 'American Express',
-  DISCOVER = 'Discover',
-  JCB = 'JCB',
-  DINERS_CLUB = 'Diners Club',
-  VISA = 'Visa',
-  MASTERCARD = 'MasterCard',
-  UNIONPAY = 'UnionPay',
-  UNKNOWN = 'Unknown',
-}
 
 export interface PaymentMethod {
   created?: number;
