+++ src/token/enum.ts
@@ -1,0 +1,12 @@
+
+export enum SourceType {
+  ThreeDeeSecure = '3ds',
+  GiroPay = 'giropay',
+  iDEAL = 'ideal',
+  SEPADebit = 'sepadebit',
+  Sofort = 'sofort',
+  AliPay = 'alipay',
+  AliPayReusable = 'alipayreusable',
+  P24 = 'p24',
+  VisaCheckout = 'visacheckout',
+}
